radio and motor test added

+++ Satellite2/Code/Components/sx1278.h
@@ -13,7 +13,7 @@ typedef enum _SX1278_MODE {
 
 typedef struct
 {
-	float frequency;
+	uint32_t frequency;
 	uint8_t power;
 	uint8_t spreadingFactor;
 	uint8_t codingRate;
